+++ sidm2/__init__.py
@@ -4,8 +4,8 @@ SID to SF2 Converter Package
 Converts Commodore 64 SID music files to SID Factory II format.
 """
 
-__version__ = "3.5.37"
-__build_date__ = "2026-05-24"
+__version__ = "3.5.38"
+__build_date__ = "2026-05-25"
 
 from .constants import *
 from .exceptions import *